+++ nemo/export/tensorrt_llm.py
@@ -28,8 +28,8 @@ import wrapt
 
 from nemo.deploy import ITritonDeployable
 from nemo.export.tarutils import TarPath, unpack_tarball
-from nemo.export.trt_llm.loader.nemo_file import get_tokenzier, load_nemo_model
-from nemo.export.trt_llm.nemo_utils import is_nemo_file, model_to_trtllm_ckpt
+from nemo.export.trt_llm.loader.nemo_file import get_tokenzier, load_nemo_model, is_nemo_file
+from nemo.export.trt_llm.converter.model_converter import model_to_trtllm_ckpt
 from nemo.export.trt_llm.qnemo import qnemo_to_tensorrt_llm
 from nemo.export.trt_llm.qnemo.tokenizer_utils import get_nmt_tokenizer
 from nemo.export.trt_llm.tensorrt_llm_build import build_and_save_engine
